+++ src/autotrade/option/analytics/forward_curve.py
@@ -9,8 +9,8 @@ from concurrent.futures import ProcessPoolExecutor, as_completed
 from tqdm.auto import tqdm
 
 """
-opt_forward_curve.py
-====================
+forward_curve.py
+================
 
 功能概述
 --------